# Performance: reduce the number of slow SQL queries in customer reports