--- a/plugins/woocommerce/changelog/performance-WOOPLUG-5382-slow-reporting-user-queries
+++ b/plugins/woocommerce/changelog/performance-WOOPLUG-5382-slow-reporting-user-queries
@@ -1,0 +1,4 @@
+Significance: patch
+Type: performance
+
+Reduce the number of slow SQL queries in customers reports.