### Submission Review Guidelines:

-   I have followed the [WooCommerce Contributing Guidelines](https://github.com/woocommerce/woocommerce/blob/trunk/.github/CONTRIBUTING.md) and the [WordPress Coding Standards](https://make.wordpress.org/core/handbook/best-practices/coding-standards/).
-   I have checked to ensure there aren't other open [Pull Requests](https://github.com/woocommerce/woocommerce/pulls) for the same update/change.
-   I have reviewed my code for [security best practices](https://developer.wordpress.org/apis/security/).
-   Following the above guidelines will result in quick merges and clear and detailed feedback when appropriate.

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### Changes proposed in this Pull Request:

Splitting #60676 into better-scoped PRs.

For the Customers vs guest report:

before: 2 queries (`0.0509 + 0.0419 = 0.0928 sec`)
<img width="2307" height="383" alt="Screenshot from 2025-09-23 08-49-59" src="https://github.com/user-attachments/assets/50926bfb-f97b-45ca-a119-7f7995f7873c" />

after: 1 query (`0.0616 sec`, 34% faster)
<img width="2313" height="217" alt="Screenshot from 2025-09-23 08-49-06" src="https://github.com/user-attachments/assets/d716f55e-b79c-4be3-a8cb-0cb2d5985073" />

For the customer list:

before: 2 queries (`0.8495 + 0.8910 = 1.7405 sec`)
<img width="2308" height="493" alt="Screenshot from 2025-09-23 09-01-14" src="https://github.com/user-attachments/assets/de25c5e2-d975-41ca-b738-6ef4969a4d59" />

after: 1 query (`0.8960 sec`, 49% faster)
<img width="2305" height="273" alt="Screenshot from 2025-09-23 08-59-50" src="https://github.com/user-attachments/assets/5ac0f880-8996-4d81-b704-74f5bae4e659" />

On average, it's ~40% faster fetching for privileged users: we run a single query, but do a bit heavier matching for each record.

### How to test the changes in this Pull Request:

- Ensure to have a setup with at least 10K users and `Query monitor` plugin installed
- Navigate to WooCommerce -> Reports -> Customers
- For both `Customers vs. guests` and `Customer list` click `Go` and ensure the reports are loading and verify SQL queries are as per description.

### Testing that has already taken place:

Has already been tested in #60676, but verifying the changes and executied SQL queries will not hurt.